qemu: br: build an Ext2 root filesystem with support for capabilities
Replace the initramfs by a proper Ext2 filesystem stored in a file and
exposed as a disk device by QEMU. This filesystem is configured with
support for Linux capabilities and the tools are compiled in, too
(getcap(8), setcap(8)).

This was made to experiment with running tee-supplicant as a non-root
user, expecting some permissions issues especially with networking
(privileged ports). It turns out that the GlobalPlatform TEE Sockets API
is client only, so no need to bind to privileged ports...

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>