toolchain.mk: arm64: download and use prebuilt toolchain
The GCC 11.3 toolchain for Aarch64 target and Aarch64 host is now
available from Arm [1] so there is no need to build it in Buildroot
anymore. Update "make toolchain" to download and extract and update
BUILDROOT_TOOLCHAIN as well to make use of that toolchain in the build.
This brings arm64 build hosts into the same situation as x86_64 ones.

Link: https://developer.arm.com/downloads/-/arm-gnu-toolchain-downloads [1]
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>